♻️ refactor: Travel Plan 컴포넌트 리팩토링 - outlet 설정 - Place Groups 컴포넌트로 분리 - routes 수정
