+++ Front/src/pages/PlaceGroups.tsx
@@ -1,0 +1,5 @@
+const PlaceGroups = () => {
+  return (<h1>그룹 이동!!</h1>);
+};
+
+export default PlaceGroups;
